cc-mode: simplify matching of "<"
Ignore the case when writing "vector < ", if some space is already manually
removed, then it may not work as expected, so it is best not to do
manual adjustment.

Also note, when re-search-backward, it is not a mirror of
re-search-forward, as re-search-backward only matches the shorter
match, not a greedy match.